socialmedia/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-polymorphic==3.0.0
djangorestframework==3.12.1
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
